+++ src/login/redux/sagas/resetpassword/postResetPasswordSaga.js
@@ -5,10 +5,10 @@ import {
   failRequest,
   putCsrfToken
 } from "../../../../sagas/common";
-import { postEmailLinkFail, postLinkCodeFail } from "../../actions/postResetPasswordActions";
+import { postEmailLinkFail } from "../../actions/postResetPasswordActions";
 import { history } from "../../../components/App/App";
 import { countDownStart } from "../../../components/LoginApp/ResetPassword/CountDownTimer";
-import * as app_actions from "../../../components/App/App_actions";
+import { loadingData, loadingDataComplete } from "../../actions/loadingDataActions";
 
 export function fetchConfigResetPassword(config, data) {
   return window
@@ -27,8 +27,10 @@ export function* postEmailLink() {
       email: state.resetPassword.email_address,
       csrf_token: state.config.csrf_token
     };
+    yield put(loadingData());
     const resp = yield call(fetchConfigResetPassword, state.config, data);
     yield put(putCsrfToken(resp));
+    yield put(loadingDataComplete());
     yield put(resp);
     if (resp.type === "POST_RESET_PASSWORD_SUCCESS") {
       history.push(`/reset-password/email-link-sent`);
@@ -36,35 +38,6 @@ export function* postEmailLink() {
     }
   } catch (error) {
     yield* failRequest(error, postEmailLinkFail(error));
-  }
-}
-
-export function fetchConfigLinkCode(config, data) {
-  return window
-    .fetch(PASSWORD_SERVICE_URL + "/verify-email/", {
-      ...postRequest,
-      body: JSON.stringify(data)
-    })
-    .then(checkStatus)
-    .then(response => response.json());
-}
-
-export function* useLinkCode() {
-  try {
-    const state = yield select(state => state);
-    if(state.resetPassword.code){
-    const data = {
-      email_code: state.resetPassword.code,
-      csrf_token: state.config.csrf_token
-    };
-    const resp = yield call(fetchConfigLinkCode, state.config, data);
-    yield put(putCsrfToken(resp));
-    yield put(resp);
-    yield put(app_actions.appLoaded());
-      history.push(`/reset-password/`);
-    }
-  } catch (error) {
-    yield put(app_actions.appLoaded());
-    yield* failRequest(error, postLinkCodeFail(error));
+    yield put(loadingDataComplete());
   }
 }
